% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucband.R
\name{insert_size_density}
\alias{insert_size_density}
\title{Compute a probability density of insert sizes}
\usage{
insert_size_density(insert_size)
}
\arguments{
\item{insert_size}{Integer. Vector of insert sizes.}
}
\value{
The insert size density.
}
\description{
Compute a probability density of insert sizes
}
\seealso{
\code{\link{nucleosome_binding_score}}
}
