% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucband.R
\name{nucleosome_banding_score}
\alias{nucleosome_banding_score}
\title{Compute the nucleosome banding score for an insert size density}
\usage{
nucleosome_banding_score(insert_size_density)
}
\arguments{
\item{insert_size_density}{Numeric. Vector of density at each insert size.}
}
\value{
The banding score
}
\description{
Compute the nucleosome banding score for an insert size density
}
\seealso{
\code{\link{insert_size_density}}
}
